Exercises 4.21 - 4.23

+++ part4/bloglist/controllers/blogs.js
@@ -1,7 +1,6 @@
 const blogsRouter = require('express').Router()
-const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
-const User = require('../models/user')
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog
@@ -10,13 +9,9 @@ blogsRouter.get('/', async (request, response) => {
     response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
     const body = request.body
-    const decodedToken = jwt.verify(request.token, process.env.SECRET)
-    if (!decodedToken.id) {
-        return response.status(401).json({ error: 'token missing or invalid' })
-    }
-    const user = await User.findById(decodedToken.id)
+    const user = request.user
 
     const blog = new Blog({
         title: body.title,
@@ -33,9 +28,15 @@ blogsRouter.post('/', async (request, response) => {
     response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id/', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+blogsRouter.delete('/:id/', middleware.userExtractor, async (request, response) => {
+    const user = request.user
+    const blog = await Blog.findById(request.params.id)
+    if( user.id === blog.user.toString() ){
+        await Blog.findByIdAndRemove(request.params.id)
+        response.status(204).end()
+    } else {
+        return response.status(401).json({ error: 'only the user who created a blog post may delete it' })
+    }
 })
 
 blogsRouter.put('/:id/', async (request, response) => {
@@ -45,7 +46,7 @@ blogsRouter.put('/:id/', async (request, response) => {
         likes: body.likes
     }
 
-    Blog.findByIdAndUpdate(request.params.id, blog, {
+    await Blog.findByIdAndUpdate(request.params.id, blog, {
         new: true,
         runValidators: true,
         context: 'query'
